ログの削除、レコーダ名をAndroid側と合わせた。その他、Warning消し残し修正。

+++ dConnectDevicePlugin/dConnectDeviceAWSIoT/dConnectDeviceAWSIoT/Classes/P2P/DPAWSIoTServerRunnable.m
@@ -47,7 +47,7 @@
         if (headerSize > 0) {
             NSData *d = [self convHeader:_headerData];
             [self.connection sendData:d.bytes length:(int)d.length];
-            [self.connection sendData:data.bytes offset:headerSize length:data.length - headerSize];
+            [self.connection sendData:data.bytes offset:headerSize length:(int) (data.length - headerSize)];
             _headerEndFlag = YES;
         }
     } else {
